Move psf module to source_catalog
